Move MeasurementGate, measure, and measure_each out of common_gates.py - The file has become too large, and we are splitting it up.
