+++ cirq/ops/measurement_gate_test.py
@@ -1,0 +1,171 @@
+# Copyright 2018 The Cirq Developers
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import numpy as np
+import pytest
+
+import cirq
+
+
+@pytest.mark.parametrize('num_qubits', [1, 2, 4])
+def test_measure_init(num_qubits):
+    assert cirq.MeasurementGate(num_qubits).num_qubits() == num_qubits
+    assert cirq.MeasurementGate(num_qubits, key='a').key == 'a'
+    assert cirq.MeasurementGate(num_qubits,
+                                invert_mask=(True,)).invert_mask == (True,)
+    assert cirq.qid_shape(cirq.MeasurementGate(num_qubits)) == (2,) * num_qubits
+    assert cirq.qid_shape(cirq.MeasurementGate(3, qid_shape=(1, 2,
+                                                             3))) == (1, 2, 3)
+    assert cirq.qid_shape(cirq.MeasurementGate(qid_shape=(1, 2, 3))) == (1, 2,
+                                                                         3)
+    with pytest.raises(ValueError, match='len.* >'):
+        cirq.MeasurementGate(5, invert_mask=(True,) * 6)
+    with pytest.raises(ValueError, match='len.* !='):
+        cirq.MeasurementGate(5, qid_shape=(1, 2))
+    with pytest.raises(ValueError, match='Specify either'):
+        cirq.MeasurementGate()
+
+
+def test_measurement_eq():
+    eq = cirq.testing.EqualsTester()
+    eq.make_equality_group(lambda: cirq.MeasurementGate(
+        1, ''), lambda: cirq.MeasurementGate(1, '', invert_mask=()), lambda:
+                           cirq.MeasurementGate(1, '', qid_shape=(2,)))
+    eq.add_equality_group(cirq.MeasurementGate(1, 'a'))
+    eq.add_equality_group(cirq.MeasurementGate(1, 'a', invert_mask=(True,)))
+    eq.add_equality_group(cirq.MeasurementGate(1, 'a', invert_mask=(False,)))
+    eq.add_equality_group(cirq.MeasurementGate(1, 'b'))
+    eq.add_equality_group(cirq.MeasurementGate(2, 'a'))
+    eq.add_equality_group(cirq.MeasurementGate(2, ''))
+    eq.add_equality_group(cirq.MeasurementGate(3, 'a'),
+                          cirq.MeasurementGate(3, 'a', qid_shape=(2, 2, 2)))
+    eq.add_equality_group(cirq.MeasurementGate(3, 'a', qid_shape=(1, 2, 3)))
+
+
+def test_measurement_full_invert_mask():
+    assert cirq.MeasurementGate(1, 'a').full_invert_mask() == (False,)
+    assert (cirq.MeasurementGate(
+        2, 'a', invert_mask=(False, True)).full_invert_mask() == (False, True))
+    assert (cirq.MeasurementGate(
+        2, 'a', invert_mask=(True,)).full_invert_mask() == (True, False))
+
+
+def test_qudit_measure_qasm():
+    assert cirq.qasm(cirq.measure(cirq.LineQid(0, 3), key='a'),
+                     args=cirq.QasmArgs(),
+                     default='not implemented') == 'not implemented'
+
+
+def test_measurement_gate_diagram():
+    # Shows key.
+    assert cirq.circuit_diagram_info(
+        cirq.MeasurementGate(1)) == cirq.CircuitDiagramInfo(("M('')",))
+    assert cirq.circuit_diagram_info(cirq.MeasurementGate(
+        1, key='test')) == cirq.CircuitDiagramInfo(("M('test')",))
+
+    # Uses known qubit count.
+    assert cirq.circuit_diagram_info(
+        cirq.MeasurementGate(3),
+        cirq.CircuitDiagramInfoArgs(known_qubits=None,
+                                    known_qubit_count=3,
+                                    use_unicode_characters=True,
+                                    precision=None,
+                                    qubit_map=None)) == cirq.CircuitDiagramInfo(
+                                        ("M('')", 'M', 'M'))
+
+    # Shows invert mask.
+    assert cirq.circuit_diagram_info(
+        cirq.MeasurementGate(2, invert_mask=(False,
+                                             True))) == cirq.CircuitDiagramInfo(
+                                                 ("M('')", "!M"))
+
+    # Omits key when it is the default.
+    a = cirq.NamedQubit('a')
+    b = cirq.NamedQubit('b')
+    cirq.testing.assert_has_diagram(cirq.Circuit(cirq.measure(a, b)), """
+a: ───M───
+      │
+b: ───M───
+""")
+    cirq.testing.assert_has_diagram(
+        cirq.Circuit(cirq.measure(a, b, invert_mask=(True,))), """
+a: ───!M───
+      │
+b: ───M────
+""")
+    cirq.testing.assert_has_diagram(
+        cirq.Circuit(cirq.measure(a, b, key='test')), """
+a: ───M('test')───
+      │
+b: ───M───────────
+""")
+
+
+def test_measurement_channel():
+    np.testing.assert_allclose(
+        cirq.channel(cirq.MeasurementGate(1)),
+        (np.array([[1, 0], [0, 0]]), np.array([[0, 0], [0, 1]])))
+    # yapf: disable
+    np.testing.assert_allclose(
+            cirq.channel(cirq.MeasurementGate(2)),
+            (np.array([[1, 0, 0, 0],
+                       [0, 0, 0, 0],
+                       [0, 0, 0, 0],
+                       [0, 0, 0, 0]]),
+             np.array([[0, 0, 0, 0],
+                       [0, 1, 0, 0],
+                       [0, 0, 0, 0],
+                       [0, 0, 0, 0]]),
+             np.array([[0, 0, 0, 0],
+                       [0, 0, 0, 0],
+                       [0, 0, 1, 0],
+                       [0, 0, 0, 0]]),
+             np.array([[0, 0, 0, 0],
+                       [0, 0, 0, 0],
+                       [0, 0, 0, 0],
+                       [0, 0, 0, 1]])))
+    np.testing.assert_allclose(
+            cirq.channel(cirq.MeasurementGate(2, qid_shape=(2, 3))),
+            (np.diag([1, 0, 0, 0, 0, 0]),
+             np.diag([0, 1, 0, 0, 0, 0]),
+             np.diag([0, 0, 1, 0, 0, 0]),
+             np.diag([0, 0, 0, 1, 0, 0]),
+             np.diag([0, 0, 0, 0, 1, 0]),
+             np.diag([0, 0, 0, 0, 0, 1])))
+    # yapf: enable
+
+
+def test_measurement_qubit_count_vs_mask_length():
+    a = cirq.NamedQubit('a')
+    b = cirq.NamedQubit('b')
+    c = cirq.NamedQubit('c')
+
+    _ = cirq.MeasurementGate(num_qubits=1, invert_mask=(True,)).on(a)
+    _ = cirq.MeasurementGate(num_qubits=2, invert_mask=(True, False)).on(a, b)
+    _ = cirq.MeasurementGate(num_qubits=3,
+                             invert_mask=(True, False, True)).on(a, b, c)
+    with pytest.raises(ValueError):
+        _ = cirq.MeasurementGate(num_qubits=1, invert_mask=(True, False)).on(a)
+    with pytest.raises(ValueError):
+        _ = cirq.MeasurementGate(num_qubits=3,
+                                 invert_mask=(True, False, True)).on(a, b)
+
+
+def test_consistent_protocols():
+    for n in range(1, 5):
+        gate = cirq.MeasurementGate(num_qubits=n)
+        cirq.testing.assert_implements_consistent_protocols(gate)
+
+        gate = cirq.MeasurementGate(num_qubits=n, qid_shape=(3,) * n)
+        cirq.testing.assert_implements_consistent_protocols(gate)
